// src/app/(sidebar-group)/statistics-dashboard/Components/types.ts
// types.ts
export interface SessionData {
  chatbotAPIRequests: number[];
  totalChatbotAPIRequests: number;
  averageChatbotAPIRequests: number;
  chatbotUsedTime: number[];
  totalChatbotUsedTime: number;
}

export interface Chatbot {
  _id: string;
  name: string;
  type: string;
  isActive: boolean;
  analyticsEnabled: boolean;
  version: number;
  createdAt: string;
  updatedAt: string;
}

export interface User {
  _id: string;
  username: string;
  email: string;
  role: string;
}

export interface StatisticsDataItem {
  _id: string;
  chatbotId: string;
  userId: string;
  data: {
    [sessionId: string]: SessionData;
  };
  createdAt: string;
  updatedAt: string;
  chatbot: Chatbot;
  user: User;
}

export interface ChatbotStat {
  id: string;
  name: string;
  type: string;
  isActive: boolean;
  creator: string;
  sessions: number;
  requests: number;
  usageTime: number;
  uniqueUsers: number;
}

export interface ChartData {
  name: string;
  value: number;
  percentage: number;
  isOthers?: boolean;
  items?: ChartData[];
}
;

export interface SessionDisplayData {
  id: string;
  chatbotName: string;
  creator: string;
  requests: number;
  usageTime: number;
}

export interface ProcessedStats {
  uniqueChatbots: number;
  uniqueCreators: number;
  totalSessions: number;
  totalRequests: number;
  totalUsageTime: number;
  chatbotStats: ChatbotStat[];
  chatbotDistributionData: ChartData[];
  sessionsData: SessionDisplayData[];
}